\name{OptionPrice-package}
\alias{OptionPrice-package}
\alias{OptionPrice}
\docType{package}
\title{
       Calculation of option price
}
\description{
       This function will use the binary tree pricing method to calculate the value of options for you. This function can calculate the value of European call option, American call option, European put option and American put option.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OptionPrice}
\packageIndices{OptionPrice}

}
\author{
\packageAuthor{OptionPrice}

Maintainer: \packageMaintainer{OptionPrice}
}
\references{
       Nothing
}

\keyword{ package }
\seealso{
       Nothing
}
\examples{
       Nothing
}
